improved local pre-aggregation of measurement data (take the average, if there is more than one sample per interval)

+++ nephele/nephele-streaming/src/main/java/eu/stratosphere/nephele/streaming/listeners/StreamListener.java
@@ -118,7 +118,7 @@ public final class StreamListener {
 			new ChannelThroughput(sourceChannelID, throughput));
 	}
 
-	public void reportBufferLatency(final ChannelID sourceChannelID, final int bufferLatency) {
+	public void reportBufferLatency(final ChannelID sourceChannelID, final double bufferLatency) {
 		this.listenerContext.getProfilingReporter().addToNextReport(
 			new OutputBufferLatency(sourceChannelID, bufferLatency));
 	}
